Renamed aux.py because Windows doesn't allow this file name
