+++ seisbench/util/__init__.py
@@ -1,5 +1,5 @@
 from .annotations import Detection, Pick
-from .aux import in_notebook
+from .auxiliary import in_notebook
 from .decorators import log_lifecycle
 from .file import (
     callback_if_uncached,
