+++ seisbench/util/auxiliary.py
